Add files via upload
This is the knwl.js webscraper that retrieves email addresses and phone numbers from the Canddi website. I have found that knwl is very poor at detecting phone numbers unless they are in a specific format, it will find numbers in this format '07463374614' but not formats such as '+44 (0) 161 237 9949' or even as '0161 237 9949'.